requirements.txt
# Telegram
telethon==1.34.0

# OpenAI
openai==1.12.0

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0

# Scheduler
apscheduler==3.10.4

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# Utilities
pytz==2024.1
aiofiles==23.2.1
httpx==0.26.0

# Logging
colorlog==6.8.2
